Add files via upload
The project entails the analysis of customer data for patterns to gain insights relating to the bike purchase behavior of customers across the demographic, financial, and behavioral accounts of individuals, thus leading to an understanding of the factors influencing purchase decisions. In addition, Excel features such as data organization and pivot tables, as well as dashboards, are used in this project to showcase results.